rewrited some cod accordint to teacher commits
