+++ myapp/src/components/ChatList/index.jsx
@@ -12,7 +12,7 @@ export default function ChatList({ data }) {
             {data.map((chat) => {
                 return (
                     <>
-                        <Link key={chat.id} to={`/chats/chat${chat.id}`}>
+                        <Link key={chat.id} to={`/chat${chat.id}`}>
                             <ListItem alignItems="flex-start" sx={{ borderTop: "solid 1px" }}>
                                 <ListItemAvatar>
                                     <Avatar alt={chat.name} src="/static/images/avatar/1.jpg" />
@@ -27,12 +27,3 @@ export default function ChatList({ data }) {
         </List>
     )
 }
-
-{/* <List sx={{ width: '100%', maxWidth: 360, bgcolor: 'background.paper' }}>
-<ListItem alignItems="flex-start">
-  <ListItemAvatar>
-    <Avatar alt="Remy Sharp" src="/static/images/avatar/1.jpg" />
-  </ListItemAvatar>
-  <ListItemText primary="Brunch this weekend?" />
-</ListItem>
-</List> */}
