+++ myapp/src/components/MessageList/style.scss
@@ -1,17 +1,6 @@
-.MessageItem{
-    background-color: aliceblue;
-    margin: 30px 10px;
-    border-radius: 10px;
+.Message {
+    color: brown;
+}
+#MessageDisplay {
     display: flex;
-    flex-direction: column;
-    text-align: start;
-        &__author{
-            color: black;
-            margin-left: 15px;
-            margin-top: 10px;
-            margin-bottom: 10px;
-        }
-        &__text{
-            margin: 0 10px;
-        }
 }
